requirements.txt
aiohttp==3.9.1
psycopg2-binary==2.9.9
